Address PR #682 review: parser extraction, Windows CI, stale-scan hint
- Extract parseNetshNetworks and localWifiNetwork to a non-build-tagged
  file (wifi_scan_netsh.go) so its tests run under the standard Ubuntu
  go test ./... PR CI job instead of being silently skipped by the
  //go:build windows tag.

- Add a Windows cross-compile step to .github/workflows/go-tests.yml
  that builds ./internal/cli/... ./cmd/wendy/ and test-compiles
  ./internal/cli/commands so Windows-only code is at least
  type-checked in PR CI. The agent is intentionally excluded — it uses
  Linux-only syscalls and only the CLI ships to Windows.

- Document on scanLocalWifiNetworks that netsh reads the Windows WLAN
  service's *cached* scan list and that a proper WlanScan API
  implementation is deferred. Introduce wifiScanCacheHint, set on
  Windows and empty on darwin/linux, and append it to the no-networks
  error and notice paths in wifi.go so Windows users see a retry hint.
