+++ go/internal/cli/commands/wifi_scan_netsh_test.go
@@ -1,5 +1,3 @@
-//go:build windows
-
 package commands
 
 import (
