+++ go/internal/cli/commands/wifi_scan_windows.go
@@ -3,21 +3,24 @@
 package commands
 
 import (
-	"bufio"
 	"fmt"
 	"os/exec"
-	"regexp"
-	"strconv"
-	"strings"
 )
 
-type localWifiNetwork struct {
-	SSID           string
-	SignalStrength int32 // 0–100 percentage, or 0 if unknown
-}
-
-// scanLocalWifiNetworks uses `netsh wlan show networks mode=bssid` to list
-// WiFi networks visible to the host machine.
+// scanLocalWifiNetworks lists WiFi networks visible to the host using
+// `netsh wlan show networks mode=bssid`.
+//
+// Important: netsh reads the Windows WLAN service's *cached* scan list.
+// The service asks the driver to rescan periodically and, per Microsoft's
+// docs, may not scan at all while already associated to a network. This
+// command therefore does not guarantee a fresh, on-demand scan the way the
+// macOS CoreWLAN and Linux `nmcli device wifi rescan` paths do.
+//
+// A proper fix is to call Native Wifi's WlanScan API
+// (https://learn.microsoft.com/en-us/windows/win32/api/wlanapi/nf-wlanapi-wlanscan)
+// via syscall — that work is intentionally deferred while this implementation
+// is the near-term fallback. Callers should treat empty results as "possibly
+// stale cache" and surface the retry hint defined by wifiScanCacheHint.
 func scanLocalWifiNetworks() ([]localWifiNetwork, error) {
 	cmd := exec.Command("netsh", "wlan", "show", "networks", "mode=bssid")
 	output, err := cmd.Output()
@@ -27,64 +30,9 @@ func scanLocalWifiNetworks() ([]localWifiNetwork, error) {
 	return parseNetshNetworks(string(output)), nil
 }
 
-// ssidLine matches `SSID 1 : MyNetwork`. It is anchored with a leading
-// boundary that rejects `BSSID` (which would otherwise match `^.SSID`).
-var ssidLine = regexp.MustCompile(`^SSID\s+\d+\s*:\s*(.*)$`)
-
-// signalLine matches `         Signal             : 78%`.
-var signalLine = regexp.MustCompile(`^\s*Signal\s*:\s*(\d+)%`)
-
-// parseNetshNetworks parses the localized text output of
-// `netsh wlan show networks mode=bssid` into a deduplicated list of SSIDs,
-// each annotated with the strongest signal strength observed across all of
-// its BSSIDs. Hidden networks (empty SSID after the colon) are skipped.
-func parseNetshNetworks(output string) []localWifiNetwork {
-	type entry struct {
-		index  int
-		signal int32
-	}
-	entries := make(map[string]*entry)
-	order := 0
-
-	var currentSSID string
-	var haveSSID bool
-
-	scanner := bufio.NewScanner(strings.NewReader(output))
-	for scanner.Scan() {
-		line := scanner.Text()
-
-		if m := ssidLine.FindStringSubmatch(line); m != nil {
-			currentSSID = strings.TrimSpace(m[1])
-			haveSSID = currentSSID != ""
-			if haveSSID {
-				if _, ok := entries[currentSSID]; !ok {
-					entries[currentSSID] = &entry{index: order}
-					order++
-				}
-			}
-			continue
-		}
-
-		if !haveSSID {
-			continue
-		}
-
-		if m := signalLine.FindStringSubmatch(line); m != nil {
-			if v, err := strconv.Atoi(m[1]); err == nil {
-				signal := int32(v)
-				if e := entries[currentSSID]; e != nil && signal > e.signal {
-					e.signal = signal
-				}
-			}
-		}
-	}
-
-	out := make([]localWifiNetwork, len(entries))
-	for ssid, e := range entries {
-		out[e.index] = localWifiNetwork{SSID: ssid, SignalStrength: e.signal}
-	}
-	return out
-}
+// wifiScanCacheHint is appended to empty-scan messages on Windows because
+// netsh reads cached results — see scanLocalWifiNetworks for details.
+const wifiScanCacheHint = "Windows scans for nearby networks periodically; if your network is missing, wait a few seconds and try again, or pass --ssid to specify it directly"
 
 const supportsKeychainLookup = false
